+++ packages/brick-kit/src/core/CustomTemplates/constants.ts
@@ -6,8 +6,10 @@ import {
   RuntimeBrickElement,
   TemplateRegistry,
 } from "@next-core/brick-types";
+import { RefObject } from "react";
 
-export const customTemplateRegistry: TemplateRegistry<CustomTemplate> = new Map();
+export const customTemplateRegistry: TemplateRegistry<CustomTemplate> =
+  new Map();
 export const appRegistered = new Set<string>();
 
 export const symbolForComputedPropsFromProxy = Symbol.for(
@@ -16,6 +18,9 @@ export const symbolForComputedPropsFromProxy = Symbol.for(
 export const symbolForRefForProxy = Symbol.for("tpl.refForProxy");
 export const symbolForParentTemplate = Symbol.for("tpl.parentTemplate");
 export const symbolForTplContextId = Symbol.for("tpl.contextId");
+export const symbolForParentRefForUseBrickInPortal = Symbol.for(
+  "parentRefForUseBrickInPortal"
+);
 
 export interface RuntimeBrickConfWithTplSymbols extends RuntimeBrickConf {
   [symbolForComputedPropsFromProxy]?: Record<string, any>;
@@ -26,4 +31,5 @@ export interface RuntimeBrickConfWithTplSymbols extends RuntimeBrickConf {
 
 export interface RuntimeBrickElementWithTplSymbols extends RuntimeBrickElement {
   [symbolForParentTemplate]?: RuntimeBrickElementWithTplSymbols;
+  [symbolForParentRefForUseBrickInPortal]?: RefObject<HTMLElement>;
 }