+++ packages/brick-kit/src/core/CustomTemplates/registerCustomTemplate.ts
@@ -36,9 +36,9 @@ export function registerCustomTemplate(
   // Collect defined properties of the template.
   const props = Object.keys(tplConstructor.proxy?.properties || {});
 
-  const nativeProp = props.some((prop) => prop in HTMLElement.prototype);
+  const nativeProp = props.find((prop) => prop in HTMLElement.prototype);
   // istanbul ignore if
-  if (nativeProp) {
+  if (nativeProp !== undefined) {
     // eslint-disable-next-line no-console
     console.error(
       `In custom template "${tagName}", "${nativeProp}" is a native HTMLElement property, and should be avoid to be used as a brick property.`