+++ packages/dev-dependencies/patches/migrateHusky.js
@@ -4,6 +4,8 @@ const execa = require("execa");
 const { readJson, writeJsonFile } = require("../utils");
 
 async function migrateHusky() {
+  const rootPackageJsonPath = path.resolve("package.json");
+  const rootPackageJson = readJson(rootPackageJsonPath);
   const huskyrcPath = path.resolve(".huskyrc");
   if (fs.existsSync(huskyrcPath)) {
     const huskyrc = readJson(huskyrcPath);
@@ -13,11 +15,9 @@ async function migrateHusky() {
     }
   }
 
-  const packageJsonPath = path.resolve("package.json");
-  const packageJson = readJson(packageJsonPath);
-  if (!packageJson.scripts["lint-staged"]) {
-    packageJson.scripts["lint-staged"] = "lint-staged";
-    writeJsonFile(packageJsonPath, packageJson);
+  if (!rootPackageJson.scripts["lint-staged"]) {
+    rootPackageJson.scripts["lint-staged"] = "lint-staged";
+    writeJsonFile(rootPackageJsonPath, rootPackageJson);
   }
 
   await execa("npx", ["husky-init"], {