+++ packages/webpack-config-factory/src/dll/webpack.dll.factory.js
@@ -1,5 +1,6 @@
 const path = require("path");
 const changeCase = require("change-case");
+const { CleanWebpackPlugin } = require("clean-webpack-plugin");
 const NextDllPlugin = require("./NextDllPlugin");
 const NextDllReferencePlugin = require("./NextDllReferencePlugin");
 const NextHashedModuleIdsPlugin = require("./NextHashedModuleIdsPlugin");
@@ -36,9 +37,12 @@ module.exports = () => {
       [changeCase.pascalCase(filename)]: Object.keys(dependencies),
     },
     output: {
-      filename: `${filename}.js`,
+      filename: isProd
+        ? `${filename}.[contenthash].js`
+        : `${filename}.bundle.js`,
       path: distPath,
       library: "[name]",
+      hashDigestLength: 8,
     },
     module: {
       rules: [
@@ -50,6 +54,7 @@ module.exports = () => {
       ],
     },
     plugins: [
+      new CleanWebpackPlugin(),
       ...dllReferences,
       new NextDllPlugin({
         name: "[name]",