+++ packages/brick-icons/CHANGELOG.md
@@ -3,13 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [2.15.2](https://github.com/easyops-cn/next-core/compare/@next-core/brick-icons@2.15.1...@next-core/brick-icons@2.15.2) (2021-05-08)
+# [2.17.0](https://github.com/easyops-cn/next-core/compare/@next-core/brick-icons@2.16.0...@next-core/brick-icons@2.17.0) (2021-05-19)
+
+### Features
+
+- add easy-website icon ([2683bb6](https://github.com/easyops-cn/next-core/commit/2683bb6f9cfa5275d806d3b0db7bd1b8f860e579))
+
+# [2.16.0](https://github.com/easyops-cn/next-core/compare/@next-core/brick-icons@2.15.3...@next-core/brick-icons@2.16.0) (2021-05-18)
+
+### Bug Fixes
+
+- refine icons loading ([b4430b3](https://github.com/easyops-cn/next-core/commit/b4430b3c081b86188d0259d777c29a15293de41f))
+
+### Features
+
+- support to async load icons and refine dll js name ([43b783b](https://github.com/easyops-cn/next-core/commit/43b783baaf570eaaf711860dceab8da913abc5f0))
+
+## [2.15.3](https://github.com/easyops-cn/next-core/compare/@next-core/brick-icons@2.15.2...@next-core/brick-icons@2.15.3) (2021-05-08)
 
 **Note:** Version bump only for package @next-core/brick-icons
 
+## [2.15.2](https://github.com/easyops-cn/next-core/compare/@next-core/brick-icons@2.15.1...@next-core/brick-icons@2.15.2) (2021-05-08)
 
-
-
+**Note:** Version bump only for package @next-core/brick-icons
 
 ## [2.15.1](https://github.com/easyops-cn/next-core/compare/@next-core/brick-icons@2.15.0...@next-core/brick-icons@2.15.1) (2021-04-26)